Modified overall UI Design
