+++ components/LoginScreen/InputView.tsx
@@ -1,37 +1,43 @@
 import React, { useState } from "react";
-import { View, TextInput, Text, TouchableOpacity, StyleSheet } from "react-native"
+import { View, TextInput, Text, StyleSheet } from "react-native"
+import FlatButton from "../FlatButton";
 
 
 interface inputProps {
     userData: (data: {userID: string, userPassword: string}[]) => void;
     propData: {userID: string, userPassword: string}[];
 }
-export default function  InputComponents (props: inputProps) {
 
+export default function  InputView (props: inputProps) {
+
+    // Takes UserName Input Value
     const [userName, setUserName] = useState("");
     const inputUserHandler = (input: string) => {
       setUserName(input);
     };
   
+    // Takes Password Input Value
     const [password, setPassword] = useState("");
     const inputPasswordHandler = (input: string) => {
       setPassword(input);
     };
   
+    // Saves both UserName & Password in form of object into an array
     const [userDetail, setUserDetail] = useState(props.propData);
     const loginHandler = () => {
       setUserDetail(
         (detail) => (detail = [{ userID: userName, userPassword: password }])
       );
+    //   Sends data to App.tsx file
       props.userData(userDetail)
     };
   
     return (
-        <View style={{ alignItems: "center", width: '100%'}}>
+        <View style={{flex:1, alignItems: "center", justifyContent: 'center', width: '100%'}}>
         <TextInput
           style={styles.inputContainer}
           placeholder="Enter Username"
-          placeholderTextColor={"#d3d3d3"}
+          placeholderTextColor={"black"}
           textAlign="center"
           onChangeText={inputUserHandler}
         ></TextInput>
@@ -39,20 +45,15 @@ export default function  InputComponents (props: inputProps) {
         <TextInput
           style={styles.inputContainer}
           placeholder="Enter Password"
-          placeholderTextColor={"#d3d3d3"}
+          placeholderTextColor={"black"}
           secureTextEntry
           textAlign="center"
           onChangeText={inputPasswordHandler}
         ></TextInput>
 
-        <View>
-          <TouchableOpacity
-            style={styles.buttonContainer}
-            onPress={loginHandler}
-          >
-            <Text style={styles.buttonText}>Login</Text>
-          </TouchableOpacity>
-        </View>
+        <FlatButton buttonText='Login' buttonFunction={loginHandler}></FlatButton>
+        <Text>or</Text>
+        <FlatButton buttonText='Sign Up' buttonFunction={()=>{}}></FlatButton>
         </View>
     )
 }
@@ -60,31 +61,16 @@ export default function  InputComponents (props: inputProps) {
 const styles = StyleSheet.create({
     
   inputContainer: {
-    borderColor: "white",
-    borderWidth: 2,
+    // borderColor: "#0d67b5",
+    // borderWidth: 2,
     borderRadius: 10,
     width: '80%',
-    height: 60,
+    height: 50,
     margin: 10,
     padding: 10,
     fontSize: 14,
+    color: 'white',
     fontStyle: "italic",
-    color: "white",
+    backgroundColor: "#eef0e9",
   },
-
-  buttonText: {
-    color: "white",
-    fontSize: 20,
-    fontWeight: "bold",
-  },
-
-  buttonContainer: {
-    backgroundColor: "#e33636",
-    padding: 10,
-    width: 100,
-    alignItems: "center",
-    margin: 10,
-    borderRadius: 10,
-  },
-
 })
