+++ package.json
@@ -10,12 +10,16 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.11",
+    "@react-navigation/native-stack": "^6.7.0",
     "expo": "~45.0.0",
     "expo-status-bar": "~1.3.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
-    "react-native-web": "0.17.7"
+    "react-native-web": "0.17.7",
+    "react-native-screens": "~3.11.1",
+    "react-native-safe-area-context": "4.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
